backend/agents/requirements.txt
conductor-python==1.2.0b1
requests==2.31.0
python-decouple==3.8
PyGithub==1.59.1
pydantic==2.4.2
google-generativeai==0.3.2
flask==3.0.0